moved fetching of certificate from handlers. Certificate is now an input to DPAPI - send()
