+++ Digipost.Api.Client.Testklient/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="digipost-api-client-shared" version="1.0.5564.24199" targetFramework="net45" />
+</packages>
